* move card component from componentLibrary to sharedComponents folder since it's not a part of DSCO
