+++ apps/src/sharedComponents/card/Card.story.tsx
@@ -2,18 +2,17 @@ import type {Meta, StoryObj} from '@storybook/react';
 import React from 'react';
 
 import {buttonColors, LinkButton} from '@cdo/apps/componentLibrary/button';
-import {CardActions} from '@cdo/apps/componentLibrary/card/CardActions';
-import {CardContent} from '@cdo/apps/componentLibrary/card/CardContent';
-import {CardHeader} from '@cdo/apps/componentLibrary/card/CardHeader';
+import {CardActions} from '@cdo/apps/sharedComponents/card/CardActions';
+import {CardContent} from '@cdo/apps/sharedComponents/card/CardContent';
+import {CardHeader} from '@cdo/apps/sharedComponents/card/CardHeader';
 import i18n from '@cdo/locale';
 
 import {Card} from './Card';
 
-import cardStyles from '@cdo/apps/componentLibrary/card/Card/card.module.scss';
+import cardStyles from '@cdo/apps/sharedComponents/card/Card/card.module.scss';
 import styles from '@cdo/apps/simpleSignUp/link-account.module.scss';
 
 const meta: Meta<typeof Card> = {
-  title: 'DesignSystem/Card', // eslint-disable-line storybook/no-title-property-in-meta
   component: Card,
 };
