+++ apps/src/sharedComponents/card/Card/card.module.scss
