+++ apps/src/simpleSignUp/lti/link/LtiLinkAccountPage/cards/LtiExistingAccountCard/index.tsx
@@ -2,15 +2,15 @@ import classNames from 'classnames';
 import React, {useContext} from 'react';
 
 import {buttonColors, Button} from '@cdo/apps/componentLibrary/button';
+import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
+import {PLATFORMS} from '@cdo/apps/lib/util/AnalyticsConstants';
+import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
 import {
   Card,
   CardActions,
   CardContent,
   CardHeader,
-} from '@cdo/apps/componentLibrary/card';
-import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
-import {PLATFORMS} from '@cdo/apps/lib/util/AnalyticsConstants';
-import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
+} from '@cdo/apps/sharedComponents/card';
 import {navigateToHref} from '@cdo/apps/utils';
 import i18n from '@cdo/locale';
