+++ apps/src/simpleSignUp/lti/link/LtiLinkAccountPage/cards/LtiNewAccountCard/index.tsx
@@ -2,16 +2,16 @@ import classNames from 'classnames';
 import React, {useContext, useRef, useState} from 'react';
 
 import {Button, buttonColors} from '@cdo/apps/componentLibrary/button';
+import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
+import {PLATFORMS} from '@cdo/apps/lib/util/AnalyticsConstants';
+import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
+import RailsAuthenticityToken from '@cdo/apps/lib/util/RailsAuthenticityToken';
 import {
   Card,
   CardActions,
   CardContent,
   CardHeader,
-} from '@cdo/apps/componentLibrary/card';
-import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
-import {PLATFORMS} from '@cdo/apps/lib/util/AnalyticsConstants';
-import analyticsReporter from '@cdo/apps/lib/util/AnalyticsReporter';
-import RailsAuthenticityToken from '@cdo/apps/lib/util/RailsAuthenticityToken';
+} from '@cdo/apps/sharedComponents/card';
 import {getAuthenticityToken} from '@cdo/apps/util/AuthenticityTokenStore';
 import i18n from '@cdo/locale';
